Refactor memcache backend into multiple files
