+++ src/backends/mod.rs
@@ -1,1 +1,3 @@
+#[cfg_attr(docsrs, doc(cfg(feature = "memcache")))]
+#[cfg(feature = "memcache")]
 pub mod memcache;
